Alignment ops

TimeDag.left — Function

left(x, y[, alignment::Alignment; initial_values=nothing])

Construct a node that ticks according to alignment with the latest value of x.

It is "left", in the sense of picking the left-hand of the two arguments x and y.

TimeDag.right — Function

right(x, y[, alignment::Alignment; initial_values=nothing])

Construct a node that ticks according to alignment with the latest value of y.

It is "right", in the sense of picking the right-hand of the two arguments x and y.

TimeDag.align — Function

align(x, y)

Form a node that ticks with the values of x whenever y ticks.

TimeDag.coalign — Function

coalign(node_1, [node_2...; alignment::Alignment]) -> Node...

Given at least one node(s) x, or values that are convertible to nodes, align all of them.

We guarantee that all nodes that are returned will have the same alignment. The values of each node will correspond to the values of the input nodes.

The choice of alignment is controlled by alignment, which defaults to UNION.

TimeDag.lag — Function

lag(x::Node, n::Integer)

Construct a node which takes values from x, but lags them by n knots.

This means that we do not introduce any new timestamps that do not appear in x, however we will not emit knots for the first n values that appear when evaluating x.