Arithmetic

Base.abs — Function

abs(x::Node)

Obtain a node with values constructed by applying abs to each input value.

Base.exp — Function

exp(x::Node)

Obtain a node with values constructed by applying exp to each input value.

Base.log — Function

log(x::Node)

Obtain a node with values constructed by applying log to each input value.

Base.log10 — Function

log10(x::Node)

Obtain a node with values constructed by applying log10 to each input value.

Base.log2 — Function

log2(x::Node)

Obtain a node with values constructed by applying log2 to each input value.

Base.sqrt — Function

sqrt(x::Node)

Obtain a node with values constructed by applying sqrt to each input value.

Base.Math.cbrt — Function

cbrt(x::Node)

Obtain a node with values constructed by applying cbrt to each input value.

Base.sign — Function

sign(x::Node)

Obtain a node with values constructed by applying sign to each input value.

Base.tan — Function

tan(x::Node)

Obtain a node with values constructed by applying tan to each input value.

Base.sin — Function

sin(x::Node)

Obtain a node with values constructed by applying sin to each input value.

Base.cos — Function

cos(x::Node)

Obtain a node with values constructed by applying cos to each input value.

Base.atan — Function

atan(x::Node)

Obtain a node with values constructed by applying atan to each input value.

Base.asin — Function

asin(x::Node)

Obtain a node with values constructed by applying asin to each input value.

Base.acos — Function

acos(x::Node)

Obtain a node with values constructed by applying acos to each input value.

Base.tanh — Function

tanh(x::Node)

Obtain a node with values constructed by applying tanh to each input value.

Base.sinh — Function

sinh(x::Node)

Obtain a node with values constructed by applying sinh to each input value.

Base.cosh — Function

cosh(x::Node)

Obtain a node with values constructed by applying cosh to each input value.

Base.atanh — Function

atanh(x::Node)

Obtain a node with values constructed by applying atanh to each input value.

Base.asinh — Function

asinh(x::Node)

Obtain a node with values constructed by applying asinh to each input value.

Base.acosh — Function

acosh(x::Node)

Obtain a node with values constructed by applying acosh to each input value.

Base.inv — Function

inv(x::Node)

Obtain a node with values constructed by applying inv to each input value.

Base.:+ — Function

+(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying + to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.:- — Function

-(x::Node)

Obtain a node with values constructed by applying - to each input value.

-(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying - to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.:* — Function

*(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying * to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.:/ — Function

/(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying / to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.:^ — Function

^(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying ^ to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.min — Function

min(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying min to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.max — Function

max(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying max to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.:! — Function

!(x::Node)

Obtain a node with values constructed by applying ! to each input value.

Base.:> — Function

>(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying > to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.:< — Function

<(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying < to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.:>= — Function

>=(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying >= to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.

Base.:<= — Function

<=(x, y[, alignment=DEFAULT_ALIGNMENT; kwargs...])

Obtain a node with values constructed by applying <= to the input values.

An alignment can optionally be specified. x and y should be nodes, or constants that can be converted to nodes.

Other keyword arguments are passed to apply.