Creating operations

Sometimes the operations contained in TimeDag will not be sufficient. This document explains how new operations can be created.

TimeDag.wrap — Function

wrap(f::Function)

Return a callable object that acts on nodes, and returns a node.

It is assumed that f is stateless and time-independent. We also assume that we will always emit a knot when the alignment semantics say we should — thus f must always return a valid output value.

If the object is called with more than one node, alignment will be performed. In this case, the final argument can be an Alignment instance, otherwise DEFAULT_ALIGNMENT will be used.

Internally this will call TimeDag.apply(f, args...); see there for further details.

TimeDag.wrapb — Function

wrapb(f::Function)

Like wrap(f), however f will be broadcasted over all input values.