Internals

This page documents internal, and less commonly-used API. Some of it will be useful for users for more advanced use-cases, like running graphs in production systems. Other parts shouldn't need to be regularly interacted with, but can be useful to know about.

Identity map

One of the key features of TimeDag.jl is avoiding duplicating work.

This is primarily achieved by ensuring that we never construct the 'same' node twice. By 'same', here we mean two nodes that we can prove will always give the same output.

One easy-to-handle case is that of a node that has identical parents & op to another. To avoid this, TimeDag.jl maintains a global identity map, of type TimeDag.IdentityMap.

Currently the only implementation of the identity map is TimeDag.WeakIdentityMap. This contains weak references to nodes, to ensure that we know about all nodes that currently exist, but don't unnecessarily prevent nodes from being garbage collected when we no longer want them.

In practice, all nodes should be constructed indirectly using TimeDag.obtain_node. This will query the global_identity_map(), and only construct a new node if an equivalent one does not already exist.

TimeDag.IdentityMap — Type

IdentityMap

An abstract identity map.

Any implementation of this type needs to implement obtain_node!.

TimeDag.obtain_node! — Function

obtain_node!(id_map::IdentityMap, parents::NTuple{N,Node}, op::NodeOp) -> Node

If a node with parents and op doesn't exist inside id_map, create and insert it.

Return either the new or existing node.

TimeDag.global_identity_map — Function

global_identity_map() -> IdentityMap

Get the global IdentityMap instance used in TimeDag.

TimeDag.WeakNode — Type

WeakNode(parents, op)

Represent a node-like object that doesn't hold strong references to its parents.

This exists purely such that hash and == do allow multiple instances of WeakNode to compare equal if they have the same parents and op.

TimeDag.WeakIdentityMap — Type

WeakIdentityMap

Represent a collection of nodes which doesn't hold strong references to any nodes.

This is useful, as it allows the existence of this cache to be somewhat transparent to the user, and they only have to care about holding on to references for nodes that they care about.

This structure contains nodes, but also node weak nodes – these allow us to determine whether we ought to create a given node.

Advanced evaluation

Sometimes evaluate does not provide enough control. This section goes into more detail about how evaluation works, and in particular how this can be useful for live systems.

Evaluation state

Batching

Scheduling

TimeDag.EvaluationState — Type

EvaluationState

All state necessary for the evaluation of some nodes.

This should be created with start_at.

Fields

ordered_node_to_children::OrderedDict{Node,Set{Node}}: a map from every node which we need to run, to its children. The ordering of the elements is such that, if evaluated in this order, all dependencies will be evaluated before they are required.

node_to_state::IdDict{Node,NodeEvaluationState}: maintain the state for every node being evaluated.

current_time::DateTime: the time to which this state corresponds.

evaluated_node_to_blocks::IdDict{Node,Vector{Block}}: the output blocks that we care about.

TimeDag.NodeEvaluationState — Type

abstract type NodeEvaluationState end

Represents any and all state that a node must retain between evaluating batches.

Instances of subtypes of NodeEvaluationState are given to run_node!.

TimeDag.start_at — Function

start_at(nodes, time_start::DateTime) -> EvaluationState

Create a sufficient EvaluationState for the evaluation of nodes.

Internally, this will determine the subgraph that needs evaluating, i.e. all the ancestors of nodes, and create a NodeEvaluationState for each of these.

TimeDag.get_up_to! — Function

get_up_to!(state::EvaluationState, time_end::DateTime)

Update the evaluation state by performing the evalution for each node.

TimeDag.run_node! — Function

run_node!(
    state::NodeEvaluationState,
    op::NodeOp{T},
    time_start::DateTime,
    time_end::DateTime,
    input_blocks::Block...
) -> Block{T}

Evaluate the given node from time_start to time_end, with the initial state. Zero or more blocks will be passed as an input; these correspond to the parents of a node, and are passed in the same order as that returned by parents(node).

We return a new Block of output knots from this node.

Warning

The implementer of run_node! must ensure that no future peeking occurs. That is, that no output knot is dependent on input knots that occur subsequently.