Miscellaneous

Random

Base.rand — Function

rand([rng=MersenneTwister(),] alignment::Node[, S, dims...])

Generate random numbers aligned to alignment, with the given rng if provided.

Semantics are otherwise very similar to the usual Base.rand:

If specified, S will be the element type, and will default to Float64 otherwise. If specified, dims should be a tuple or vararg of integers representing the dimensions of an array.

NB The values of alignment will be ignored.

Tip

If provided, rng will be copied before it is used. This is to ensure reproducability when evaluating a node multiple times.

Conditional

TimeDag.zap_missing — Function

zap_missing(x::Node)

Obtain a node which ticks with the values of x, so long as that value is not missing.

The value_type of the node that is returned will always be the nonmissingtype of the value_type of x.

In the case that x cannot tick with missing (based on its value_type), we just return x.